try add vipseg r50 configs

+++ dvis/ctvis.py
@@ -353,10 +353,10 @@ class CTMinVIS(nn.Module):
 
         # merge outputs
         outputs = {}
-        outputs['pred_logits'] = torch.cat([x['pred_logits'] for x in out_list], dim=1).detach().cpu()
-        outputs['pred_masks'] = torch.cat([x['pred_masks'] for x in out_list], dim=2).detach().cpu()
-        outputs['pred_embds'] = torch.cat([x['pred_embds'] for x in out_list], dim=2).detach().cpu()
-        outputs['pred_reid_embed'] = torch.cat([x['pred_reid_embed'] for x in out_list], dim=2).detach().cpu()
+        outputs['pred_logits'] = torch.cat([x['pred_logits'] for x in out_list], dim=1).to(torch.float32).detach().cpu()
+        outputs['pred_masks'] = torch.cat([x['pred_masks'] for x in out_list], dim=2).to(torch.float32).detach().cpu()
+        outputs['pred_embds'] = torch.cat([x['pred_embds'] for x in out_list], dim=2).to(torch.float32).detach().cpu()
+        outputs['pred_reid_embed'] = torch.cat([x['pred_reid_embed'] for x in out_list], dim=2).to(torch.float32).detach().cpu()
 
         return outputs
